apps/api/pyproject.toml
[project]
name = "medai-council-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.15",
    "openai>=1.57.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.9.0",
    "pypdf>=6.10.2",
    "python-multipart>=0.0.26",
    "numpy>=2.4.4",
]
